move platform to model
