+++ src/consts.py
@@ -30,40 +30,11 @@ class BITNESS(enum.Enum):
     B32 = 32
 
 
-class HP(enum.Enum):
-    """HumbleBundle platform code name shown in subproducts>download section"""
-    WINDOWS = 'windows'
-    MAC = 'mac'
-    LINUX = 'linux'
-    ANDROID = 'android'
-    AUDIO = 'audio'
-    EBOOK = 'ebook'
-    ASMJS = 'asmjs'
-    UNITYWASM = 'unitywasm'
-    VIDEO = 'video'
-    COMEDY = 'comedy'
-
-    def __eq__(self, other):
-        if type(other) == str:
-            return self.value == other
-        return super().__eq__(other)
-
-    def __hash__(self):
-        return hash(self.value)
-
-
 GAME_PLATFORMS = set([HP.WINDOWS, HP.MAC, HP.LINUX])
 NON_GAME_BUNDLE_TYPES = {'mobilebundle', 'softwarebundle', 'bookbundle', 'audiobookbundle', 'comicsbundle', 'rpgbookbundle', 'mangabundle'}
 
-if sys.platform == 'win32':
-    CURRENT_SYSTEM = HP.WINDOWS
-elif sys.platform == 'darwin':
-    CURRENT_SYSTEM = HP.MAC
-else:
-    raise PlatformNotSupported('GOG Galaxy 2.0 supports only Windows and macos for now')
-
-IS_WINDOWS = HP.WINDOWS == CURRENT_SYSTEM
-IS_MAC = HP.MAC == CURRENT_SYSTEM
+IS_WINDOWS = sys.platform == 'win32'
+IS_MAC = sys.platform == 'darwin'
 
 if platform.machine().endswith('64'):
     CURRENT_BITNESS = BITNESS.B64
