+++ src/local/pathfinder.py
@@ -9,8 +9,8 @@ from consts import HP
 
 
 class PathFinder:
-    def __init__(self, system: HP):
-        self.system = system
+    def __init__(self, is_windows: bool):
+        self.is_windows = is_windows
 
     def find_executables(self, path: Union[str, PurePath]) -> List[str]:
         folder = Path(path)
@@ -26,7 +26,7 @@ class PathFinder:
         return execs
 
     def is_exe(self, path: str) -> bool:
-        if self.system == HP.WINDOWS:
+        if self.is_windows:
             return path.endswith('.exe')
         else:
             return os.access(path, os.X_OK)
