+++ src/model/platform.py
@@ -1,0 +1,23 @@
+import enum
+
+
+class HP(enum.Enum):
+    """HumbleBundle platform code name shown in subproducts>download section"""
+    WINDOWS = 'windows'
+    MAC = 'mac'
+    LINUX = 'linux'
+    ANDROID = 'android'
+    AUDIO = 'audio'
+    EBOOK = 'ebook'
+    ASMJS = 'asmjs'
+    UNITYWASM = 'unitywasm'
+    VIDEO = 'video'
+    COMEDY = 'comedy'
+
+    def __eq__(self, other):
+        if type(other) == str:
+            return self.value == other
+        return super().__eq__(other)
+
+    def __hash__(self):
+        return hash(self.value)
